Move Brunch architecture checks to app tests
