+++ apps/brunch-agent/docs/task-dependencies.json
@@ -40,7 +40,10 @@
     ],
     "test:unit": [
       "@apps/brunch-agent#build",
+      "@hashintel/brunch-agent#build",
       "@hashintel/brunch-agent-binding-flue#build",
+      "@hashintel/brunch-agent-plugin-gherkin#build",
+      "@hashintel/brunch-agent-plugin-sdcpn#build",
       "@hashintel/brunch-agent-transport-aisdk#build",
       "@hashintel/petrinaut-core#build"
     ]
